split config tests to separate file
